Add SLF4J integration as a new library
Forwards log requests to SLF4J, which in turn allows to use Logback as a
backend. This provides rich appenders and output configuration via
`logback.xml` files. It can replace the default `ConsoleLogger` on
JVM and Android platforms. To keep things simple, the JavaScript and
Apple platform implementation log to the console.